working redux
I created a) a store constant b) a reducer/actions for the userData (to have the userName and userId accessible). c) wrapped the tutorial in a <Provider> to access the store. d) Rn, when clicking next, an action is dispatched to change the username. This is then the state used for all other tutorial screens